Move usMaxContext calculation to its own file
